Fix: Changed structure, to create store on a separated file making easier to follow the data
